+++ src/auth/auth.module.ts
@@ -5,14 +5,14 @@ import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { JwtRefreshStrategy } from './strategies/jwt-refresh.strategy';
 import { JwtModule } from '@nestjs/jwt';
-import { UserRepository } from 'src/repositories/user.repository';
+import { UserRepository } from 'src/user/repositories/user.repository';
 import { JwtAccessStrategy } from './strategies/jwt-access.strategy';
 import { JwtAccessNicknameCheckStrategy } from './strategies/jwt-access-nickname-check.strategy';
 import { UserService } from 'src/user/user.service';
 import { UserModule } from 'src/user/user.module';
-import { PreferenceRepository } from 'src/repositories/preference.repository';
-import { FolderRepository } from 'src/repositories/folder.repository';
-import { FolderPlaceRepository } from 'src/repositories/folder-place.repository';
+import { PreferenceRepository } from 'src/user/repositories/preference.repository';
+import { FolderRepository } from 'src/folder/repositories/folder.repository';
+import { FolderPlaceRepository } from 'src/folder/repositories/folder-place.repository';
 
 @Global()
 @Module({
