+++ src/auth/strategies/jwt-access-nickname-check.strategy.ts
@@ -1,7 +1,7 @@
 import { ForbiddenException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, ExtractJwt } from 'passport-jwt';
-import { UserRepository } from 'src/repositories/user.repository';
+import { UserRepository } from 'src/user/repositories/user.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 
 export class JwtAccessNicknameCheckStrategy extends PassportStrategy(
