+++ src/folder/folder.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { FolderController } from './folder.controller';
 import { FolderService } from './folder.service';
-import { FolderRepository } from 'src/repositories/folder.repository';
-import { FolderPlaceRepository } from 'src/repositories/folder-place.repository';
+import { FolderRepository } from 'src/folder/repositories/folder.repository';
+import { FolderPlaceRepository } from 'src/folder/repositories/folder-place.repository';
 
 @Module({
   controllers: [FolderController],
