+++ src/place/place.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { PlaceController } from './place.controller';
 import { PlaceService } from './place.service';
-import { PlaceRepository } from 'src/repositories/place.repository';
+import { PlaceRepository } from 'src/place/repositories/place.repository';
 import { PlaceTagRepository } from 'src/repositories/place-tag.repository';
-import { PlaceImageRepository } from 'src/repositories/place-image.repository';
+import { PlaceImageRepository } from 'src/place/repositories/place-image.repository';
 import { PlaceDetailRepository } from 'src/repositories/place-detail.repository';
 import { TagModule } from 'src/tag/tag.module';
-import { S3Service } from 'src/s3/s3.service';
+import { S3Service } from 'src/place/s3.service';
 
 @Module({
   controllers: [PlaceController],
