+++ src/place/repositories/place-image.repository.ts
