+++ src/place/s3.service.ts
