+++ src/trip/repositories/trip.repository.ts
