+++ src/trip/trip.module.ts
@@ -6,7 +6,7 @@ import { Trip } from 'src/entities/trip.entity';
 import { TripStyle } from 'src/entities/trip-style.entity';
 import { AccommodationSchedule } from 'src/entities/accomodation-schedule.entity';
 import { PlaceSchedule } from 'src/entities/place-schedule.entity';
-import { TripRepository } from 'src/repositories/trip.repository';
+import { TripRepository } from 'src/trip/repositories/trip.repository';
 
 @Module({
   imports: [
